requirements.txt
PyYAML>=6.0
dataclasses-json>=0.5.7